requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
certifi==2023.7.22
cffi==1.15.1
cfgv==3.4.0
charset-normalizer==3.2.0
cryptography==41.0.3
defusedxml==0.7.1
Deprecated==1.2.14
diff-match-patch==20230430
distlib==0.3.7
Django==4.1.5
django-braces==1.15.0
django-cors-headers==3.13.0
django-filter==23.2
django-import-export==3.3.3
django-jalali-date==1.0.2
djangorestframework==3.14.0
et-xmlfile==1.1.0
Faker==16.6.0
filelock==3.13.1
identify==2.5.32
idna==3.4
jdatetime==4.1.1
jwcrypto==1.5.0
MarkupPy==1.14
nodeenv==1.8.0
odfpy==1.4.1
openpyxl==3.1.2
Pillow==9.4.0
platformdirs==4.0.0
pre-commit==3.5.0
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python-decouple==3.7
python3-openid==3.2.0
pytz==2022.7.1
PyYAML==6.0.1
redis==4.4.2
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
tablib==3.5.0
urllib3==2.0.4
virtualenv==20.24.7
wrapt==1.15.0
xlrd==2.0.1
xlwt==1.3.0
